api/requirements.txt
django>=1.10
djangorestframework
django-environ
https://github.com/chibisov/drf-extensions/archive/master.tar.gz
